+++ CHANGELOG.md
@@ -5,7 +5,8 @@
 * Rename `--watch-and-serve` to `--server`
 * Update `MarkdownRenderer` defaults for `descriptive_class_title` and
   `descriptive_class_title`
-* Added `MkdocsRenderer.content_directory` (defaults to `content`)
+* Added `MkdocsRenderer.content_directory` (defaults to `content`, used to be
+  `docs` before)
 * Renamed `MkdocsRenderer.clean_docs_directory_on_render` to `.clean_render`
   and make it default to `False`
 * Fixed `--server` option, now reloads the `PydocMarkdown` config, does no